+++ crates/sncast/tests/e2e/account/create.rs
@@ -19,7 +19,7 @@ use test_case::test_case;
 
 #[test_case("oz"; "oz_account_type")]
 #[test_case("argent"; "argent_account_type")]
-// TODO(#3118): Re-enable this test once braavos integration is restored
+// TODO(#3357)
 // #[test_case("braavos"; "braavos_account_type")]
 #[tokio::test]
 pub async fn test_happy_case(account_type: &str) {
@@ -332,7 +332,7 @@ pub async fn test_account_already_exists() {
 
 #[test_case("oz"; "oz_account_type")]
 #[test_case("argent"; "argent_account_type")]
-// TODO(#3118)
+// TODO(#3357)
 // #[test_case("braavos"; "braavos_account_type")]
 #[tokio::test]
 pub async fn test_happy_case_keystore(account_type: &str) {
@@ -822,6 +822,7 @@ fn get_keystore_account_pattern(account_type: AccountType, class_hash: Option<&s
     to_string_pretty(&account_json).unwrap()
 }
 
+// TODO(#3357): Remove this test when Braavos account creation is enabled
 #[test]
 fn test_braavos_disabled() {
     let temp_dir = tempdir().expect("Unable to create a temporary directory");
@@ -849,8 +850,7 @@ fn test_braavos_disabled() {
         output,
         indoc! {r"
         command: account create
-        error: Using Braavos accounts is temporarily disabled because they don't yet work with starknet 0.13.5.
-            Visit this link to read more: https://community.starknet.io/t/starknet-devtools-for-0-13-5/115495#p-2359168-braavos-compatibility-issues-3
+        error: Creating Braavos accounts is currently disabled. Please use the Braavos web wallet to create an account and later import it with sncast
         "},
     );
 }